requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Machine Learning
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
Pillow==10.1.0
opencv-python==4.8.1.78

# Data processing
pytz==2023.3
python-dateutil==2.8.2

# Web scraping / screenshot capture
selenium==4.15.2
webdriver-manager==4.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
